+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### BC breaks
 
+- The ReadModelTestCase is renamed to SerializableReadModelTestCase
 - We moved the Doctrine DBAL event store implementation to a [separate repository](https://github.com/broadway/event-store-dbal)
 - We moved the Elasticsearch read model implementation to a [separate repository](https://github.com/broadway/read-model-elasticsearch) 
 - We moved the Symfony bundle to a [separate repository](https://github.com/broadway/broadway-bundle)